+++ docs/conf.py
@@ -73,10 +73,17 @@ extensions = [
     "matplotlib.sphinxext.plot_directive",
     "m2r2",
     "nbsphinx",
-    "sphinx.ext.autodoc.typehints"
+    "sphinx.ext.autodoc.typehints",
 ]
 
 # autodoc options
+
+autodoc_type_aliases = {}
+
+# this does not work properly yet
+autodoc_typehints = "none"
+autodoc_typehints_format = "short"
+autodoc_typehints_description_target = "documented"
 
 autodoc_mock_imports = [
     "System",
@@ -333,8 +340,10 @@ def linkcode_resolve(domain, info):
 
 # extlinks
 
+
 extlinks = {
-    "rhino": ("https://developer.rhino3d.com/api/RhinoCommon/html/T_%s.htm", "%s")
+    "rhino": ("https://developer.rhino3d.com/api/RhinoCommon/html/T_%s.htm", "%s"),
+    "blender": ("https://docs.blender.org/api/2.93/%s.html", "%s")
 }
 
 # -- Options for HTML output ----------------------------------------------
@@ -348,7 +357,7 @@ html_theme_options = {
     "package_old_versions_txt": "https://compas.dev/compas/doc_versions.txt"
 }
 html_context = {}
-html_static_path = []
+html_static_path = sphinx_compas_theme.get_html_static_path()
 html_extra_path = []
 html_last_updated_fmt = ""
 html_copy_source = False