+++ src/compas/artists/__init__.py
@@ -5,6 +5,14 @@ artists
 
 .. currentmodule:: compas.artists
 
+.. rst-class:: lead
+
+For visualization of data objects such as geometry objects, robots, and data structures, COMPAS provides "artists".
+Every data object type is paired with a corresponding artist type that is capable of visualizing the data.
+This package provides base artist classes with pluggable methods
+that receive an implementation from plugins defined by various visualization contexts.
+
+
 Classes
 =======
 
@@ -30,6 +38,18 @@ Exceptions
 
     DataArtistNotRegistered
 
+
+Pluggables
+==========
+
+.. autosummary::
+    :toctree: generated/
+    :nosignatures:
+
+    clear
+    redraw
+    register_artists
+
 """
 from __future__ import print_function
 from __future__ import absolute_import
@@ -43,6 +63,10 @@ from .primitiveartist import PrimitiveArtist
 from .robotmodelartist import RobotModelArtist
 from .shapeartist import ShapeArtist
 from .volmeshartist import VolMeshArtist
+
+from .artist import clear  # noqa: F401
+from .artist import redraw  # noqa: F401
+from .artist import register_artists  # noqa: F401
 
 BaseRobotModelArtist = RobotModelArtist