+++ src/compas/artists/artist.py
@@ -63,6 +63,15 @@ def _get_artist_cls(data, **kwargs):
 
 class Artist(object):
     """Base class for all artists.
+
+    Class Attributes
+    ----------------
+    AVAILABLE_CONTEXTS : list[str]
+        The available visualization contexts.
+    CONTEXT : str or None
+        The current visualization context is one of :attr:`AVAILABLE_CONTEXTS`.
+    ITEM_ARTIST : dict[str, dict[Type[:class:`compas.data.Data`], Type[:class:`compas.artists.Artist`]]]
+        Dictionary mapping data types to the corresponding artists types per visualization context.
     """
 
     __ARTISTS_REGISTERED = False
@@ -92,8 +101,8 @@ class Artist(object):
         Returns
         -------
         :class:`compas.artists.Artist`
-            An artist of the type matching the provided item according to the item-artist map ``~Artist.ITEM_ARTIST``.
-            The map is created by registering item-artist type pairs using ``~Artist.register``.
+            An artist of the type matching the provided item according to the item-artist map :attr:`~Artist.ITEM_ARTIST`.
+            The map is created by registering item-artist type pairs using :meth:`~Artist.register`.
         """
         artist_type = _get_artist_cls(item, **kwargs)
         artist = artist_type(item, **kwargs)
@@ -101,25 +110,68 @@ class Artist(object):
 
     @staticmethod
     def build_as(item, artist_type, **kwargs):
+        """Build an artist with the given type.
+
+        Parameters
+        ----------
+        artist_type : :class:`compas.artists.Artist`
+        **kwargs : dict[str, Any], optional
+            The keyword arguments (kwargs) collected in a dict.
+            For relevant options, see the parameter lists of the matching artist type.
+
+        Returns
+        -------
+        :class:`compas.artists.Artist`
+            An artist of the given type.
+        """
         artist = artist_type(item, **kwargs)
         return artist
 
     @staticmethod
     def clear():
+        """Clear all objects from the view.
+
+        Returns
+        -------
+        None
+        """
         return clear()
 
     @staticmethod
     def redraw():
+        """Redraw the view.
+
+        Returns
+        -------
+        None
+        """
         return redraw()
 
     @staticmethod
     def register(item_type, artist_type, context=None):
+        """Register an artist type to a data type.
+
+        Parameters
+        ----------
+        item_type : :class:`compas.data.Data`
+            The type of data item.
+        artist_type : :class:`compas.artists.Artist`
+            The type of the corresponding/compatible artist.
+        context : {'Rhino', 'Grasshopper', 'Blender', 'Plotter'}, optional
+            The visualization context in which the pair should be registered.
+
+        Returns
+        -------
+        None
+        """
         Artist.ITEM_ARTIST[context][item_type] = artist_type
 
     @abstractmethod
     def draw(self):
+        """The main drawing method."""
         raise NotImplementedError
 
     @staticmethod
     def draw_collection(collection):
+        """Drawing method for drawing an entire collection of objects."""
         raise NotImplementedError