+++ src/compas/files/__init__.py
@@ -5,6 +5,11 @@ files
 
 .. currentmodule:: compas.files
 
+.. rst-class:: lead
+
+This package provides classes for working with selected file formats that are capable
+of storing information about 2D and 3D geometry, robots, pointclouds, ...
+
 
 DXF
 ===