+++ src/compas/geometry/__init__.py
@@ -900,16 +900,6 @@ from .trimesh import (  # noqa: E402
 if not compas.IPY:
     from .icp import icp_numpy
 
-from .collections import (  # noqa: E402
-    Collection,
-    PointCollection,
-)
-if not compas.IPY:
-    from .collections import (
-        CollectionNumpy,
-        PointCollectionNumpy
-    )
-
 from .curves import (
     Curve,
     NurbsCurve
@@ -1220,9 +1210,6 @@ __all__ = [
     'Transformation',
     'Translation',
 
-    'Collection',
-    'PointCollection',
-
     'Curve',
     'NurbsCurve',
 
@@ -1252,6 +1239,4 @@ if not compas.IPY:
         'local_to_world_coordinates_numpy',
         'delaunay_from_points_numpy',
         'voronoi_from_points_numpy',
-        'CollectionNumpy',
-        'PointCollectionNumpy',
     ]