+++ src/compas_blender/geometry/__init__.py
@@ -5,21 +5,23 @@ geometry
 
 .. currentmodule:: compas_blender.geometry
 
-Object-oriented convenience wrappers for native Blender geometry.
+Classes
+=======
 
 .. autosummary::
     :toctree: generated/
+    :nosignatures:
 
-    BlenderCurve
-    BlenderMesh
+
+Plugins
+=======
+
+.. autosummary::
+    :toctree: generated/
+    :nosignatures:
+
+    booleans.boolean_difference_mesh_mesh
+    booleans.boolean_intersection_mesh_mesh
+    booleans.boolean_union_mesh_mesh
 
 """
-from ._geometry import BaseBlenderGeometry
-from .curve import BlenderCurve
-from .mesh import BlenderMesh
-
-__all__ = [
-    'BaseBlenderGeometry',
-    'BlenderCurve',
-    'BlenderMesh'
-]