+++ src/compas_ghpython/artists/networkartist.py
@@ -20,6 +20,10 @@ class NetworkArtist(GHArtist, NetworkArtist):
     ----------
     network : :class:`compas.datastructures.Network`
         A COMPAS network.
+    **kwargs : dict, optional
+        Additional keyword arguments.
+        See :class:`compas_ghpython.artists.GHArtist` and :class:`compas.artists.NetworkArtist` for more info.
+
     """
 
     def __init__(self, network, **kwargs):
@@ -30,26 +34,30 @@ class NetworkArtist(GHArtist, NetworkArtist):
 
         Returns
         -------
-        tuple
-            list of :class:`Rhino.Geometry.Point3d` and list of :class:`Rhino.Geometry.Line`
+        list[:rhino:`Rhino.Geometry.Point3d`]
+            The objects representing the nodes of the network.
+        list[:rhino:`Rhino.Geometry.Line`]
+            The objects representing the edges of the network.
+
         """
-        return (self.draw_nodes(), self.draw_edges())
+        return self.draw_nodes(), self.draw_edges()
 
     def draw_nodes(self, nodes=None, color=None):
         """Draw a selection of nodes.
 
         Parameters
         ----------
-        nodes: list, optional
+        nodes: list[hashable], optional
             The selection of nodes that should be drawn.
-            Default is ``None``, in which case all nodes are drawn.
-        color: 3-tuple or dict of 3-tuple, optional
+            Default is None, in which case all nodes are drawn.
+        color: tuple[int, int, int] or dict[hashable, tuple[int, int, int]], optional
             The color specification for the nodes.
-            The default color is ``(255, 255, 255)``.
+            The default color is :attr:`NetworkArtist.default_nodecolor`.
 
         Returns
         -------
-        list of :class:`Rhino.Geometry.Point3d`
+        list[:rhino:`Rhino.Geometry.Point3d`]
+
         """
         self.node_color = color
         node_xyz = self.node_xyz
@@ -68,16 +76,17 @@ class NetworkArtist(GHArtist, NetworkArtist):
 
         Parameters
         ----------
-        edges : list, optional
+        edges : list[tuple[hashable, hashable]], optional
             A list of edges to draw.
-            The default is ``None``, in which case all edges are drawn.
-        color : tuple or dict of tuple, optional
+            The default is None, in which case all edges are drawn.
+        color : tuple[int, int, int] or dict[tuple[hashable, hashable], tuple[int, int, int]], optional
             The color specification for the edges.
-            The default color is the value of ``~NetworkArtist.default_edgecolor``.
+            The default color is the value of :attr:`NetworkArtist.default_edgecolor`.
 
         Returns
         -------
-        list of :class:`Rhino.Geometry.Line`
+        list[:rhino:`Rhino.Geometry.Line`]
+
         """
         self.edge_color = color
         node_xyz = self.node_xyz