+++ src/compas_plotters/plotter.py
@@ -1,3 +1,21 @@
+"""
+********************************************************************************
+plotter
+********************************************************************************
+
+.. currentmodule:: compas_plotters.plotter
+
+Classes
+=======
+
+.. autosummary::
+    :toctree: generated/
+    :nosignatures:
+
+    Plotter
+
+"""
+
 import os
 from typing import Callable, Optional, Tuple, List, Union
 from typing_extensions import Literal
@@ -19,17 +37,68 @@ class Singleton(type):
         return cls._instances[cls]
 
 
-class Plotter(metaclass=Singleton):
+class Plotter:
     """Plotter for the visualization of COMPAS geometry.
 
     Parameters
     ----------
-    view : tuple, optional
+    view : tuple[tuple[float, float], tuple[float, float]], optional
         The area of the axes that should be zoomed into view.
-        Default is ``([-10, 10], [-3, 10])``.
-    figsize : tuple, optional
-        The size of the figure in inches.
-        Default is ``(8, 5)``
+    figsize : tuple[float, float], optional
+        Size of the figure in inches.
+    dpi : float, optional
+        Resolution of the figure in "dots per inch".
+    bgcolor : tuple[float, float, float], optional
+        Background color for the figure canvas.
+    show_axes : bool, optional
+        If True, show the axes of the figure.
+    zstack : {'natural', 'zorder'}, optional
+        If ``'natural'``, the drawing elements appear in the order they were added.
+        If ``'natural'``, the drawing elements are added based on their `zorder`.
+
+    Attributes
+    ----------
+    viewbox : tuple[tuple[float, float], tuple[float, float]]
+        X min-max and Y min-max of the area of the axes that is zoomed into view.
+    axes : matplotlib.axes.Axes, read-only
+        `matplotlib` axes object used by the figure.
+        For more info, see the documentation of the Axes class ([1]_) and the axis and tick API ([2]_).
+    figure : matplotlib.figure.Figure, read-only
+        `matplotlib` figure instance.
+        For more info, see the figure API ([3]_).
+    bgcolor : tuple[float, float, float]
+        Background color of the figure.
+    title : str
+        Title of the plot.
+    artists : list[:class:`compas_plotters.artists.PlotterArtist`]
+        Artists that should be included in the plot.
+
+    Class Attributes
+    ----------------
+    fontsize : int
+        Default fontsize used by the plotter.
+
+    References
+    ----------
+    .. [1] https://matplotlib.org/api/axes_api.html
+    .. [2] https://matplotlib.org/api/axis_api.html
+    .. [3] https://matplotlib.org/api/figure_api.html
+
+    Examples
+    --------
+    >>> from compas.geometry import Point, Plane, Circle
+    >>> from compas_plotters.plotter import Plotter
+
+    Create a plotter instance.
+
+    >>> plotter = Plotter()
+
+    Add COMPAS objects.
+
+    >>> plotter.add(Point(0, 0, 0))
+    <compas_plotters.artists.pointartist.PointArtist object at 0x17880eb80>
+    >>> plotter.add(Circle(Plane.worldXY(), 1.0))
+    <compas_plotters.artists.circleartist.CircleArtist object at 0x10d136e80>
 
     """
 
@@ -55,7 +124,6 @@ class Plotter(metaclass=Singleton):
 
     @property
     def viewbox(self) -> Tuple[Tuple[float, float], Tuple[float, float]]:
-        """([xmin, xmax], [ymin, ymax]): The area of the axes that is zoomed into view."""
         return self._viewbox
 
     @viewbox.setter
@@ -67,24 +135,6 @@ class Plotter(metaclass=Singleton):
 
     @property
     def axes(self) -> matplotlib.axes.Axes:
-        """Returns the axes subplot matplotlib object.
-
-        Returns
-        -------
-        Axes
-            The matplotlib axes object.
-
-        Notes
-        -----
-        For more info, see the documentation of the Axes class ([1]_) and the
-        axis and tick API ([2]_).
-
-        References
-        ----------
-        .. [1] https://matplotlib.org/api/axes_api.html
-        .. [2] https://matplotlib.org/api/axis_api.html
-
-        """
         if not self._axes:
             figure = plt.figure(facecolor=self.bgcolor,
                                 figsize=self.figsize,
@@ -120,84 +170,27 @@ class Plotter(metaclass=Singleton):
 
     @property
     def figure(self) -> matplotlib.figure.Figure:
-        """Returns the matplotlib figure instance.
-
-        Returns
-        -------
-        Figure
-            The matplotlib figure instance.
-
-        Notes
-        -----
-        For more info, see the figure API ([1]_).
-
-        References
-        ----------
-        .. [1] https://matplotlib.org/2.0.2/api/figure_api.html
-
-        """
         return self.axes.get_figure()
 
     @property
-    def canvas(self):
-        """Returns the canvas of the figure instance.
-        """
-        return self.figure.canvas
-
-    @property
     def bgcolor(self) -> str:
-        """Returns the background color.
-
-        Returns
-        -------
-        str
-            The color as a string (hex colors).
-
-        """
         return self._bgcolor
 
     @bgcolor.setter
     def bgcolor(self, value: Union[str, Tuple[float, float, float]]):
-        """Sets the background color.
-
-        Parameters
-        ----------
-        value : str, tuple
-            The color specification for the figure background.
-            Colors should be specified in the form of a string (hex colors) or
-            as a tuple of normalized RGB components.
-
-        """
         self._bgcolor = value
         self.figure.set_facecolor(value)
 
     @property
     def title(self) -> str:
-        """Returns the title of the plot.
-
-        Returns
-        -------
-        str
-            The title of the plot.
-
-        """
         return self.figure.canvas.get_window_title()
 
     @title.setter
     def title(self, value: str):
-        """Sets the title of the plot.
-
-        Parameters
-        ----------
-        value : str
-            The title of the plot.
-
-        """
         self.figure.canvas.set_window_title(value)
 
     @property
     def artists(self) -> List[PlotterArtist]:
-        """list of :class:`compas_plotters.artists.PlotterArtist`"""
         return self._artists
 
     @artists.setter
@@ -251,61 +244,58 @@ class Plotter(metaclass=Singleton):
         self.axes.autoscale_view()
 
     def add(self,
-            item: Union[compas.geometry.Circle,
-                        compas.geometry.Ellipse,
-                        compas.geometry.Line,
-                        compas.geometry.Point,
-                        compas.geometry.Polygon,
-                        compas.geometry.Polyline,
-                        compas.geometry.Vector,
+            item: Union[compas.geometry.Primitive,
+                        compas.datastructures.Network,
                         compas.datastructures.Mesh],
             artist: Optional[PlotterArtist] = None,
             **kwargs) -> PlotterArtist:
         """Add a COMPAS geometry object or data structure to the plot.
+
+        Parameters
+        ----------
+        item
+            A COMPAS geometric primitive, network, or mesh.
+        artist
+            Type of artist to use for drawing.
+
+        Returns
+        -------
+        :class:`PlotterArtist`
+
         """
         if not artist:
             if self.zstack == 'natural':
                 zorder = 1000 + len(self._artists) * 100
-                artist = PlotterArtist(item, zorder=zorder, context='Plotter', **kwargs)
+                artist = PlotterArtist(item, plotter=self, zorder=zorder, context='Plotter', **kwargs)
             else:
-                artist = PlotterArtist(item, context='Plotter', **kwargs)
+                artist = PlotterArtist(item, plotter=self, context='Plotter', **kwargs)
         artist.draw()
         self._artists.append(artist)
         return artist
 
-    def add_as(self,
-               item: Union[compas.geometry.Circle,
-                           compas.geometry.Ellipse,
-                           compas.geometry.Line,
-                           compas.geometry.Point,
-                           compas.geometry.Polygon,
-                           compas.geometry.Polyline,
-                           compas.geometry.Vector,
-                           compas.datastructures.Mesh],
-               artist_type: PlotterArtist,
-               **kwargs) -> PlotterArtist:
-        """Add a COMPAS geometry object or data structure using a specific artist type."""
-        artist = PlotterArtist(item, artist_type=artist_type, context='Plotter', **kwargs)
-        artist.draw()
-        self._artists.append(artist)
-        return artist
+    # def add_as(self,
+    #            item: Union[compas.geometry.Primitive,
+    #                        compas.datastructures.Network,
+    #                        compas.datastructures.Mesh],
+    #            artist_type: PlotterArtist,
+    #            **kwargs) -> PlotterArtist:
+    #     """Add a COMPAS geometry object or data structure using a specific artist type."""
+    #     artist = PlotterArtist(item, artist_type=artist_type, context='Plotter', **kwargs)
+    #     artist.draw()
+    #     self._artists.append(artist)
+    #     return artist
 
     def add_from_list(self, items, **kwargs) -> List[PlotterArtist]:
         """Add multiple COMPAS geometry objects and/or data structures from a list."""
         artists = []
         for item in items:
-            artist = self.add(item, **kwargs)
+            artist = self.add(item, plotter=self, **kwargs)
             artists.append(artist)
         return artists
 
     def find(self,
-             item: Union[compas.geometry.Circle,
-                         compas.geometry.Ellipse,
-                         compas.geometry.Line,
-                         compas.geometry.Point,
-                         compas.geometry.Polygon,
-                         compas.geometry.Polyline,
-                         compas.geometry.Vector,
+             item: Union[compas.geometry.Primitive,
+                         compas.datastructures.Network,
                          compas.datastructures.Mesh]) -> PlotterArtist:
         """Find a geometry object or data structure in the plot."""
         for artist in self._artists: